requirements.txt
python-dotenv>=1.0.1
langfuse>=2.38.0
openai>=1.40.0
langchain>=0.2.16
langchain-core>=0.2.38
langchain-openai>=0.1.22
langgraph>=0.2.14

